dom/workers/test/test_workersDisabled.xul

<?xml version="1.0"?>
<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<window title="DOM Worker Threads Test"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        onload="test();">
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>

  <script type="application/javascript">
  <![CDATA[
    function test()
    {
      const enabledPref = "dom.workers.enabled";
      const message = "Hi";

      var prefs = Components.classes["@mozilla.org/preferences-service;1"]
                            .getService(Components.interfaces.nsIPrefBranch);
      is(prefs.getBoolPref(enabledPref), true, "Workers should be enabled.");

      prefs.setBoolPref(enabledPref, false);

      ok("Worker" in window, "Worker constructor should be available.");
      ok("ChromeWorker" in window,
         "ChromeWorker constructor should be available.");

      var worker = new ChromeWorker("workersDisabled_worker.js");
      worker.onmessage = function(event) {
        is(event.data, message, "Good message.");
        prefs.clearUserPref(enabledPref);
        SimpleTest.finish();
      }
      worker.postMessage(message);

      SimpleTest.waitForExplicitFinish();
    }
  ]]>
  </script>

  <body xmlns="http://www.w3.org/1999/xhtml">
    <p id="display"></p>
    <div id="content" style="display:none;"></div>
    <pre id="test"></pre>
  </body>
  <label id="test-result"/>
</window>
